+++ tests/Concerto/PanelBundle/Controller/FunctionalTests/TestNodeConnectionControllerTest.php
@@ -26,7 +26,7 @@ class TestNodeConnectionControllerTest extends AFunctionalTest
             "name" => "testFlow",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_FLOW,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));
@@ -38,7 +38,7 @@ class TestNodeConnectionControllerTest extends AFunctionalTest
             "name" => "test_s1",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_CODE,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));
@@ -50,7 +50,7 @@ class TestNodeConnectionControllerTest extends AFunctionalTest
             "name" => "test_s2",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_CODE,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));