+++ tests/Concerto/PanelBundle/Controller/FunctionalTests/TestNodePortControllerTest.php
@@ -30,7 +30,7 @@ class TestNodePortControllerTest extends AFunctionalTest
             "name" => "testFlow",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_FLOW,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));
@@ -42,7 +42,7 @@ class TestNodePortControllerTest extends AFunctionalTest
             "name" => "test_s1",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_CODE,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));