+++ tests/Concerto/PanelBundle/Controller/FunctionalTests/TestSessionLogControllerTest.php
@@ -30,7 +30,7 @@ class TestSessionLogControllerTest extends AFunctionalTest
             "name" => "test",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_CODE,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));