+++ tests/Concerto/TestBundle/Controller/FunctionalTests/TestRunnerControllerTest.php
@@ -27,7 +27,7 @@ class TestRunnerControllerTest extends AFunctionalTest {
             "name" => "test",
             "description" => "description",
             "code" => "print('start')",
-            "visibility" => Test::VISIBILITY_FEATURED,
+            "visibility" => Test::VISIBILITY_REGULAR,
             "type" => Test::TYPE_CODE,
             "accessibility" => ATopEntity::ACCESS_PUBLIC
         ));